refactor(DocThread): move FaqLayout into component dir with salon
- FaqLayout.tsx -> FaqLayout/index.tsx
- salon/faq_layout.ts -> FaqLayout/salon/index.ts
